Initial Commit
Initial Commit. Not all features are implemented. The library works fine
for basics use of simple (and most of derived) SI Unit. Some Non SI Unit
are implemented as well for convenience and proof of concept.
Offset for Temperature Missing. Candela unit definiton are missing. Unit
tests are limited for complex use of unit. Support for complex or
integral data type not tested.
VS2010 Support.