Simplification in Derived unit definition (Force, Pressure, ...)
This cause a problem in Suffix expansion. Some simplification like m*m-1
are not done compile time cause the term are on different product. Tried
to implement a template specialization but this cause a partial
specialization ambiguity. That might be not possible to implement deep
compile time simplification of unit because of that.