Added a compiler switch (NO_STATIC_UNIT_SUFFIXES_EVALUATION)
MSTest with C++\CLI is having serious problems loading static template variable and were asserting in the CLR with strange errors. I didn't investigate to get to the bottom of this but test few compiler option to see if it has impact, but I was unable to modify the strange behavior.

I decided to implements a compiler switch using a macro to be able to turn off automatic suffix evaluation at boot-time. I implemented a static method (SuffixesInitializer::Initialize()) that needs to be called before suffix string had to be evaluated using compact form.