Fix bug when converting unit with base factor other then 1. (ie. Kilogram, Metric Ton.)
While computing the scaling factor (the ratio of both quantity value),  the factor type defined in the unit definition was used, which is incorrect in that case where base unit is not aligned on the simplest symbol (like gram vs kilogram).

At that point, the library is quite stable and is performing as it was originally planned more than ten years ago.

More to go... Liter next. Hopefully, this will go through quite seamlessly.