Added Fahrenheit unit.
In my opinion, no unit library can be complete without the support of the most challenging unit of all time: the Fahrenheit. Remembering the syntax of the name itself is a challenge on its own.

Few things to note in that commit.

The use of boost::enable_if in conjunction with one trait (is_supporting_offset) to track at compile time the correct partial template to choose. This the most optimized way to handle the offset in my opinion and that prevents partial specialization ambiguities that comes from extending the previous approach that I had chosen before.

The use of one more parameters in the Convert method to scale the value in the correct unit before and after applying the offset correction. This is due mostly to the fact that both addition and multiplication are not associative when use in conjunction.

Fahrenheit is now working... that itself if one big achievement. I can say that the unit manipulation library is now working as a whole with maybe few bugs still remaining. But with Metric Ton and Fahrenheit working... the most difficult part is now behind me.